Moved ControllerConstants to controller package
