+++ src/main/java/pl/sosinski/patryk/letsmeet/web/controller/MainController.java
@@ -3,9 +3,9 @@ package pl.sosinski.patryk.letsmeet.web.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.EVENTS_URL;
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.LOGIN_URL;
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.LOGIN_VIEW;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.EVENTS_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.LOGIN_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.LOGIN_VIEW;
 
 @Controller
 public class MainController {
