+++ src/main/java/pl/sosinski/patryk/letsmeet/web/controller/ParticipantWebController.java
@@ -15,8 +15,13 @@ import pl.sosinski.patryk.letsmeet.web.model.request.ParticipantRequestModel;
 import javax.validation.Valid;
 import java.util.logging.Logger;
 
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.LOGIN_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.PARTICIPANTS_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.REGISTRATION_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.REGISTRATION_VIEW;
+
 @Controller
-@RequestMapping("/participants")
+@RequestMapping(PARTICIPANTS_URL)
 public class ParticipantWebController {
 
     private final Logger LOGGER = Logger.getLogger(ParticipantWebController.class.getName());
@@ -27,28 +32,28 @@ public class ParticipantWebController {
         this.participantManagerService = participantManagerService;
     }
 
-    @GetMapping("/registration")
+    @GetMapping(REGISTRATION_URL)
     public String showRegistrationForm(ModelMap modelMap) {
         LOGGER.info("showRegistrationForm()");
 
         modelMap.addAttribute("participant", new ParticipantRequestModel());
 
         LOGGER.info("showRegistrationForm()");
-        return "/security/registration";
+        return REGISTRATION_VIEW;
     }
 
-    @PostMapping("/registration")
+    @PostMapping(REGISTRATION_URL)
     public String registerNewParticipant(@ModelAttribute("participant") @Valid ParticipantRequestModel participantRequestModel,
                                          BindingResult bindingResult) throws EmailAlreadyExistsException {
         LOGGER.info("registerNewParticipant(" + participantRequestModel + ")");
 
         if(bindingResult.hasErrors()){
-            return "/security/registration";
+            return REGISTRATION_VIEW;
         }
 
         ParticipantModel createdParticipantModel = participantManagerService.create(participantRequestModel);
 
         LOGGER.info("registerNewParticipant(...) = " + createdParticipantModel);
-        return "redirect:/loginPage";
+        return "redirect:" + LOGIN_URL;
     }
 }
