+++ src/main/java/pl/sosinski/patryk/letsmeet/web/controller/rest/EventCategoryController.java
@@ -14,7 +14,7 @@ import pl.sosinski.patryk.letsmeet.web.model.EventCategoryModel;
 import java.util.List;
 import java.util.logging.Logger;
 
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.EVENT_CATEGORY_URI;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.EVENT_CATEGORY_URI;
 
 @RestController
 @RequestMapping(value = EVENT_CATEGORY_URI)
