+++ src/test/java/pl/sosinski/patryk/letsmeet/web/controller/MainControllerTest.java
@@ -6,13 +6,12 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.LOGIN_URL;
-import static pl.sosinski.patryk.letsmeet.web.controller.rest.ControllerConstants.LOGIN_VIEW;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.LOGIN_URL;
+import static pl.sosinski.patryk.letsmeet.web.controller.ControllerConstants.LOGIN_VIEW;
 
 @SpringBootTest
 @AutoConfigureMockMvc
